requirements.txt
numpy==1.25.2
PyYAML==6.0.1
torch==2.3.1
torch_geometric==2.5.3
torchist==0.2.3
tqdm==4.66.4
wandb==0.17.2